Store stock config, rather than actual build config, in the kernel.
When the kernel is used to boot the system from the boot partition
rather than the recovery partition, Android must be fooled into
believing that the kernel was built from the stock config. Otherwise,
Android will issue a pop-up message on boot completion and the user will
be unable to dismiss it.